requirements.txt
# Django Core
Django==4.2.8
python-decouple==3.8
Pillow==10.1.0

# Database
psycopg2-binary==2.9.9

# Django Admin
django-jazzmin==2.6.0

# REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
django-cors-headers==4.3.1

# Async Tasks
celery==5.3.4
redis==5.0.1

# WebSockets
channels==4.0.0
daphne==4.0.0

# Utilities
django-extensions==3.2.3
requests==2.31.0

# Static Files
whitenoise==6.6.0

# Production Server
gunicorn==21.2.0

# Development
ipython==8.18.1

# Validators
validate-docbr==1.10.0

# Date/Time
python-dateutil==2.8.2

# PDF Generation
reportlab==4.0.7
weasyprint==60.1

# Excel Export
openpyxl==3.1.2
xlsxwriter==3.1.9